Add very rough API usage info, and remove incomplete Opengraph info
There's been interest in embedding Bandcrash as a backend on a web service, so I figured I'd document that a bit.

And, I did some testing on the Opengraph tags and they turned out to be incomplete, and there is probably something else that needs to happen for the various iframe providers to actually work. So I've removed that section for now.