Massively cut down on pre-encode wait time
The `encode_files` step had a whole bunch of stuff that was needlessly blocking the start of the actual encode process, which in turn was blocking the GUI.

All I/O-bound metadata is now gathered only when it's needed, by the process that needs it.

This also removes some bogus micro-optimizations that were really just there for some edge-case tests.

This massively speeds up the encoding process, especially when gathering data from a NAS, in the most common cases.

In the uncommon case that there are multiple tracks with the same input filename, there is a little time wasted because the preview encoder needs to recompute the md5sum each time, but that's an edge case and fixing it is a micro-optimization, which is, as I like to say, a bug in disguise.